Prelim results
Now that we have some results, it seems that fixed key lengths (whilst the winner) are limiting us from performing additional rounds of compression.
We can look at optimizing the third try and creating variable length keys with the proviso that we can identify the start of a key (i.e. letter -> num [-> x]  [-> x]) etc. This may end up turning into a good alternative with an inbuilt compression, so to speak. Will return #TODO third try under development and will not run as is