Focussing down on transactions
Adding the particular nuances of the EOS transactions and how these relate to Bitcoin's UTXO model (binary Merkle Trees) and Ethereum's more in-depth light client querying capabilities which allow not only single transactions to be verified. Ethereum's Modified Patricia Merkle Tree provides light clients with the ability to query the blockchain state (account balances and so forth)

EOS transaction structure is also very interesting and this will be added next.