SWARM content can be in many formats
The content can be an index of a database or a virtual server. This may prove useful in packaging up information for sale. This is apposed to just using the file system and text files etc.
A user may be able to create a database view which can feed the low security advertising data to potential buyers. In addition to this the high security content which is actually sold can be encrypted (pulled from a different database view) before ever seeing SWARM.