Investigating Scatters boundaries
It seems that Scatter is/will be a powerful part of the holistic dApp development of not just EOS but potentially all blockchains.
This web site shows how it is used to safely change private and public keys of an already existing account on a blockchain (in this early case EOS). This is very interesting and is a huge advance over service like MetaMask in my opinion. Especially due to the fact that it could/should be on handheld devices.