Next task is to rewrite the transferFunds function
Everything is ready to have the HTML/JS app transfer some funds out of the Scatter account to the existing CyberMiles testnet account.
I just need to find out what the best practice is in relation to calling Scatter accounts in the Scatter Desktop i.e. instead of using the explicit account number, there would surely be a variable/placeholder such as account.scatter or something.